Add files via upload
MLTP (Machine Language Translation Project) is a project made by me, using Python and libraries such as NLTK, Re, and Numpy. 
The main idea of this project was to create a small part of Google Translate with only translation between 2 languages, English to Turkish and Azerbaijani to Turkish.
It uses 2 corpus that has 10000 lines of translated sentences to train itself. When the user gives a sentence to be translated, the algorithm first finds the most similar 10 sentences inside the correct corpus, then it trains itself with these 10 similar sentences that have similar words in them with NLTK library. Then the algorithm translates the sentence by the probabilities for each word, aligns them in the correct order, and prints it out.
This project depends heavily on the data set in the corpus, the higher the data set you have gives the optimal results.